update template config.

+++ src/core/program.js
@@ -13,26 +13,31 @@ utils.checkNodeVersion();
 
 program
   .version(pkg.version)
-  .command('init')
-  .description('Initialize workspace.')
-  .action(async ()=> {
-    const runtimeDir = path.resolve('.');
-    const configDir = path.resolve(__dirname, '../src/template');
+  .command('init [dir]')
+  .description('初始化工作目录.')
+  .action(async (dir)=> {
+    let inputDir = dir || '';
+    const runtimeDir = path.resolve('.', inputDir);
+    const configDir = path.resolve(__dirname, '../template');
+    if (!utils.exist(runtimeDir)) {
+      utils.error(`路径不存在，请选择正确的路径: ${runtimeDir}`);
+      return;
+    }
     if (utils.isDirEmpty(runtimeDir)) {
       fs.copySync(configDir, runtimeDir);
-      utils.info('Initialized successfully.');
+      utils.info('初始化成功，生成了config，data，middlewate等文件夹。');
     } else {
-      utils.error('Please select an empty folder.');
+      utils.error('请选择一个空的文件夹.');
     }
   });
 
 program
   .command('start [dir]')
-  .description('Start the mock server.')
+  .description('启动Mock Server.')
   .action(async (dir)=> {
     let inputDir = dir || '';
     const runtimeDir = path.resolve('.', inputDir);
-    utils.info(`Loading workspace from ${runtimeDir}.`);
+    utils.info(`从路径${runtimeDir}加载配置文件.`);
     utils.checkWorkSpace(runtimeDir);
     await startServer(runtimeDir);
